requirements.txt
torch~=2.3.1+cu118
opencv-python~=4.10.0.82
future~=1.0.0
pillow~=10.3.0
torchvision~=0.18.1
numpy~=1.26.4